# doc/openapi.yml
openapi: 3.0.1
info:
  title: ICDC Storage API
  description: >
    # Introduction
    
    The `Storage` service provides easy HTTP API to manage **Object Storage** (aka S3 buckets) and **iSCSI Block Storage** functionality of ICDC Cloud Platform.  
    Using this API you can script the complex actions that are required for your business case.  
    The API documentation starting from general overview and followed by reference information on specific endpoints.
    
    ## Object Storage
    
    **Object Storage** is a technology based on OpenSource `Ceph` project.  

    This functionality allows to create special `s3 user` accounts with the purpose of accessing buckets and store objects into the buckets.  
    Access can be provided with two protocols:
    - Amazon S3-compatible Protocol
    - Openstack Swift protocol

    ## iSCSI Block Storage

    This functionality allows to create iSCSI Disks (backed by RBD disk protocol) and iSCSI Clients (consumer devices).  
    iSCSI Clients (identified by IQN) can connect through iSCSI Portals (aka iSCSI Gateways, hosted on CloudGateway VM)
    and get access to assigned iSCSI Disks (which are RBD disk on shared `Ceph` cluster).

  version: 2.0.2
  contact:
    email: support@icdc.io
  license:
    name: Apache 2.0
    url: http://www.apache.org/licenses/LICENSE-2.0.html
servers:
  - description: Public URL
    url: https://api.{loc}.icdc.io/api/storage/v2
    variables:
      loc:
        default: dcz.lab
...
security:
  - JWT:
    # There is list of roles, not scopes:
    - operator
    - owner
    - admin
    - member
    Account: []
    Role: []
  - basicAuth: []
    Account: []
    Role: []
paths:
...
  /iscsi/clients/{client_id}/disks/{id}:
    delete:
      operationId: iscsi.clients.disks.unassign
      tags: [iSCSI]
      summary: Unassign disk from specific iSCSI client
      description: >-
        With this endpoint specific iSCSI disk can be unassigned (resigned) from the ISCSI client (aka consumer device).  
        Basically, multiple disks can be assigned to multiple clients.  
        Client and disk records are created, assigned, unassigned and deleted by end-users.  
        Client has one owner user. Disk has one owner user.  

        Effectively owner\admin can assign disk of one user to another user's client.  
        User which owns the client can unassign any disk from own client.  

        #### Role-based access
          - `member` can unassign any disks from own client.
          - `owner`\`admin` can unassign any disk from any client in authorized account
          - `operator` can unassign any disk from any client in any account
      security:
        - JWT: []
          Account: []
          Role: []
        - basicAuth: []
          Account: []
          Role: []
      parameters:
        - $ref: '#/components/parameters/ClientId'
        - in: path
          name: id
          required: true
          schema:
            type: integer
      responses:
        '204':
          description: No Content
        '401':
          $ref: '#/components/responses/Unauthorized'
        '404':
          $ref: '#/components/responses/Notfound'

  # @DS: reviewed 2024-02-02
...
  /iscsi/snapshots/{snapshot_id}/rollback:
    post:
      operationId: iscsi.snapshots.rollback
      tags: [iSCSI]
      summary: Restore iSCSI disk from snapshot
      description: >-
        Restore state of iSCSI disk from snapshot.  
        Disks and snapshots are created by end-users.  
        Restore a disk may be a time consuming operation, so we don't guarantee return in before default 30 sec UI timeout.  

        #### Role-based access
          - `member` can restore disk only from snapshots of own disk.
          - `owner`\`admin` can restore disk from any snapshot in authorized account
          - `operator` can restore disk from any snapshot in any account
      security:
        - JWT: []
          Account: []
          Role: []
        - basicAuth: []
          Account: []
          Role: []
      parameters:
        - $ref: '#/components/parameters/DiskId'
        - in: path
          name: snapshot_name
          required: true
          schema:
            type: string
      responses:
        '200':
          description: iSCSI Disk
          $ref: '#/components/schemas/iSCSI_Disk'
        '401':
          $ref: '#/components/responses/Unauthorized'
        '404':
          $ref: '#/components/responses/Notfound'

  # @DS: reviewed 2024-02-02
  # TODO: Currently not implemeted endpoint, it should replace implemented functions which use snapshot_name as id
  # TODO: Put disk into restore state,  as we do not have other notification channel and task queue.
...
components:
  responses:
...
    Unauthorized:
      description: Unauthorized
      content:
        application/json:
          schema:
            $ref: '#/components/schemas/Error'
          example:
            code: '401'
            message: Not authorized to perform operation
    Notfound:
      description: Specific object not found
      content:
        application/json:
          schema:
            $ref: '#/components/schemas/Error'
          example:
            code: '404'
            message: Not found
...
  schemas:
...
    Error:
      type: object
      properties:
        code:
          type: string
        message:
          type: string
      required:
        - code
        - message
    Pool:
      type: object
      properties:
        id:
          type: integer
          example: 9
        type:
          type: string
          example: s3
        class:
          type: string
          example: nvme
        name:
          type: string
          example: fast_pool


    # @DS: reviewed 2025-12-10
    Account_Ref:
      description: >-
        Reference to Account object, used in other objects to reference account
      type: object
      properties:
        id:
          readOnly: true
          type: integer
          example: 17
        name:
          readOnly: true
          type: string
          example: acc1
      required:
        - id
        - name
    # @DS: reviewed 2025-12-10
...
    iSCSI_Disk:
      description: >-
        Disk create in specific target
      type: object
      required:
        - name
        - size_gb
        - owner
        - pool_id
      properties:
        id:
          readOnly: true
          type: integer
          example: 66
        name:
          type: string
          example: My Disk 1
        size_gb:
          type: integer
          example: 100
        from_snapshot_id:
          writeOnly: true
          type: integer
          example: 71
        owner:
          type: string
          example: user@example.net
        pool_id:
          writeOnly: true
          type: string
          example: 4
        pool:
          readOnly: true
          oneOf:
            - $ref: '#/components/schemas/Pool'
        account_name:
          writeOnly: true
          type: string
          example: acc2
        account:
          readOnly: true
          oneOf:
            - $ref: '#/components/schemas/Account_Ref'


    # @DS: reviewed 2024-01-29
...
  parameters:
...
    DiskId:
      in: path
      name: disk_id
      description: >-
        ID of iSCSI Disk, as stored in Storage database
      required: true
      schema:
        type: integer
      example: 71
    ClientId:
      in: path
      name: client_id
      description: >-
        ID of iSCSI Client, as stored in Storage database
      required: true
      schema:
        type: integer
      example: 81
  securitySchemes:
    JWT:
      type: openIdConnect
      openIdConnectUrl: https://login.lab.icdc.io/realms/devel/.well-known/openid-configuration
    basicAuth:
        type: http
        scheme: basic
    Account:
      type: apiKey
      in: header
      #TODO: [DS] rename to x-auth-account
      name: x-auth-account
    Role:
      type: apiKey
      in: header
      #TODO: [DS] rename to x-auth-role
      name: x-auth-role